Add distributed data parallel benchmark tool
The need for this tool was motivated by #28883. In the past, we have
done ad-hoc benchmarking, but it's time for something more structured.

It would be nice to add more model architectures so that we can get a
full picture of the performance impact of a code change simply by
running this suite a few times.

ghstack-source-id: 0120497b28eb840894ab724a4e7c4a2a7ea92164
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35198